requirements.txt
anthropic
requests
feedparser
elevenlabs
flask
pillow
pyyaml
schedule
